Create components.py
