+++ zyte_common_items/__init__.py
@@ -1,13 +1,12 @@
 # flake8: noqa
-from .base import (
+from .base import Item, is_data_container
+from .components import (
     AdditionalProperty,
     AggregateRating,
     Brand,
     Breadcrumb,
     Gtin,
     Image,
-    Item,
     Metadata,
-    is_data_container,
 )
 from .items import Product, ProductVariant
