+++ zyte_common_items/base.py
@@ -126,44 +126,3 @@ class Item(_ItemBase):
             item.update({key: cls.from_list(item.get(key, [])) for key, cls in (from_list or {}).items()})
 
         return item
-
-
-@attrs.define(slots=True)
-class AdditionalProperty(Item):
-    name: str
-    value: str
-
-
-@attrs.define(slots=True, kw_only=True)
-class AggregateRating(Item):
-    bestRating: Optional[float] = None
-    ratingValue: Optional[float] = None
-    reviewCount: Optional[int] = None
-
-
-@attrs.define(slots=True)
-class Brand(Item):
-    name: str
-
-
-@attrs.define(slots=True, kw_only=True)
-class Breadcrumb(Item):
-    name: Optional[str] = None
-    url: Optional[str] = None
-
-
-@attrs.define(slots=True)
-class Gtin(Item):
-    type: str
-    value: str
-
-
-@attrs.define(slots=True, kw_only=True)
-class Image(Item):
-    url: Optional[str]
-
-
-@attrs.define(slots=True, kw_only=True)
-class Metadata(Item):
-    dateDownloaded: Optional[str] = None
-    probability: Optional[float] = None
